Remove func: getHumanChoice as it was causing err
Refactored the eventListener to outside of
getHumanChoice function as it was causing
unnecessary logic errors.

Now the eventListener calls PlayRound function
whenever a button is pressed, PlayRound function
has all the logic of rock,Paper,Scissors.
Then getScore function is called after result
which computes user and computer score. It also
calls update the score by calling the function
showScore for every compute and ends it when
one of the entities get a score of 5.