Create a message section to show player moves
Created a section to show a win/lose message
whenever a user clicks a button. Placed it at
the bottom of the webpage and styled it.

Coded this section by creating element in js file
and adding the text content and appending it
to a parent element. Also added horizontal Lines
for separation of messages.